moved SettingsTab out of Settings split general tab
